refactor(api): split contexts.py
This was getting unwieldy and a bit of a pain to edit. instrument_context.py is
still pretty big but I think this is a DX gain.
